sponsor products and updates initial refactoring and fixes
