+++ react/src/dashboard/components/pages/fund-provider/elements/BudgetFundSponsorProducts.tsx
@@ -5,7 +5,6 @@ import FundProvider from '../../../../props/models/FundProvider';
 import Organization from '../../../../props/models/Organization';
 import useFilter from '../../../../hooks/useFilter';
 import Paginator from '../../../../modules/paginator/components/Paginator';
-import Product from '../../../../props/models/Product';
 import { useOrganizationService } from '../../../../services/OrganizationService';
 import useSetProgress from '../../../../hooks/useSetProgress';
 import LoadingCard from '../../../elements/loading-card/LoadingCard';
@@ -13,8 +12,9 @@ import TableRowActions from '../../../elements/tables/TableRowActions';
 import useUpdateProduct from '../hooks/useUpdateProduct';
 import StateNavLink from '../../../../modules/state_router/StateNavLink';
 import EmptyCard from '../../../elements/empty-card/EmptyCard';
+import SponsorProduct from '../../../../props/models/Sponsor/SponsorProduct';
 
-type ProductLocal = Product & {
+type ProductLocal = SponsorProduct & {
     allowed: boolean;
 };
 
@@ -53,7 +53,10 @@ export default function BudgetFundSponsorProducts({
     );
 
     const mapProduct = useCallback(
-        (product: Product) => ({ ...product, allowed: fundProvider.products.indexOf(product.id) !== -1 }),
+        (product: SponsorProduct) => ({
+            ...product,
+            allowed: fundProvider.products.indexOf(product.id) !== -1,
+        }),
         [fundProvider.products],
     );
