+++ react/src/dashboard/components/pages/fund-provider/elements/SubsidyFundProducts.tsx
@@ -5,7 +5,6 @@ import FundProvider from '../../../../props/models/FundProvider';
 import Organization from '../../../../props/models/Organization';
 import useFilter from '../../../../hooks/useFilter';
 import Paginator from '../../../../modules/paginator/components/Paginator';
-import Product, { DealHistory } from '../../../../props/models/Product';
 import useSetProgress from '../../../../hooks/useSetProgress';
 import LoadingCard from '../../../elements/loading-card/LoadingCard';
 import { useFundService } from '../../../../services/FundService';
@@ -13,8 +12,9 @@ import useUpdateProduct from '../hooks/useUpdateProduct';
 import TableRowActions from '../../../elements/tables/TableRowActions';
 import StateNavLink from '../../../../modules/state_router/StateNavLink';
 import EmptyCard from '../../../elements/empty-card/EmptyCard';
+import SponsorProduct, { DealHistory } from '../../../../props/models/Sponsor/SponsorProduct';
 
-type ProductLocal = Product & {
+type ProductLocal = SponsorProduct & {
     allowed: boolean;
     active_deal: DealHistory;
 };
@@ -46,7 +46,7 @@ export default function SubsidyFundProducts({
     );
 
     const mapProduct = useCallback(
-        (product: Product) => {
+        (product: SponsorProduct) => {
             const activeDeals = product.deals_history ? product.deals_history.filter((deal) => deal.active) : [];
 
             return {
