+++ react/src/dashboard/router/routes.tsx
@@ -19,7 +19,6 @@ import OrganizationCreate from '../components/pages/organizations-edit/Organizat
 import Auth2FA from '../components/pages/auth/Auth2FA';
 import ProviderOverview from '../components/pages/provider-overview/ProviderOverview';
 import Offices from '../components/pages/offices/Offices';
-import Products from '../components/pages/products/Products';
 import ProductsCreate from '../components/pages/products-edit/ProductsCreate';
 import ProductsEdit from '../components/pages/products-edit/ProductsEdit';
 import ProductsView from '../components/pages/products-view/ProductsView';
@@ -89,8 +88,8 @@ import SponsorFundUnsubscriptions from '../components/pages/sponsor-fund-unsubsc
 import OrganizationsContacts from '../components/pages/organizations-contacts/OrganizationsContacts';
 import Payouts from '../components/pages/payouts/Payouts';
 import PayoutsView from '../components/pages/payouts-view/PayoutsView';
-import SponsorProducts from '../components/pages/products/sponsor/SponsorProducts';
-import SponsorProductLogs from '../components/pages/products/sponsor/SponsorProductLogs';
+import ProductsViewHistory from '../components/pages/products-view/SponsorProductsView';
+import Products from '../components/pages/products/Products';
 
 const router = new RouterBuilder();
 
@@ -210,15 +209,9 @@ router.state('sponsor-provider-organizations', <SponsorProviderOrganizations />,
     fallbackState: 'organizations',
 });
 
-router.state('sponsor-products', <SponsorProducts />, {
-    path: `/organisaties/:organizationId/sponsor/products`,
-    altPath: `/organizations/:organizationId/sponsor/products`,
-    fallbackState: 'organizations',
-});
-
-router.state('sponsor-product-logs', <SponsorProductLogs />, {
-    path: `/organisaties/:organizationId/sponsor/fund/:fundId/fund_provider/:fundProviderId/product/:id`,
-    altPath: `/organizations/:organizationId/sponsor/fund/:fundId/fund_provider/:fundProviderId/product/:id`,
+router.state('sponsor-product-history', <ProductsViewHistory />, {
+    path: `/organisaties/:organizationId/producten/:id/geschiedenis`,
+    altPath: `/organizations/:organizationId/products/:id/history`,
     fallbackState: 'organizations',
 });
